move graphics code around a bit
